requirements.txt
blinker==1.9.0
click==8.3.0
contourpy==1.3.2
cycler==0.12.1
Flask==3.1.2
flask-cors==6.0.1
fonttools==4.60.1
imbalanced-learn==0.14.0
imblearn==0.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
numpy==2.2.6
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
threadpoolctl==3.6.0
tzdata==2025.2
Werkzeug==3.1.3
